Shouldn't this work?  I am unsure of the exact use case and why you need the two description variables.
It seemed the only difference was that your isChanged variable but that
doesnt seem to be used anywhere!

Either way there is a tonne of nuts stuff going on here and I will go
through and rewite it as prod code and you can use it as a reference.

bh